+++ futurehome/src/ha/mqtt_components/climate.ts
@@ -158,18 +158,6 @@ export interface ClimateComponent extends BaseComponent {
   optimistic?: boolean;
 
   /**
-   * The payload that represents the available state.
-   * Default: "online"
-   */
-  payload_available?: string;
-
-  /**
-   * The payload that represents the unavailable state.
-   * Default: "offline"
-   */
-  payload_not_available?: string;
-
-  /**
    * The payload sent to turn off the device.
    * Default: "OFF"
    */