+++ futurehome/src/ha/mqtt_components/device_tracker.ts
@@ -83,16 +83,4 @@ export interface DeviceTrackerComponent extends BaseComponent {
    * Default: '"None"'
    */
   payload_reset?: string;
-
-  /**
-   * The payload that represents the available state.
-   * Default: "online"
-   */
-  payload_available?: string;
-
-  /**
-   * The payload that represents the unavailable state.
-   * Default: "offline"
-   */
-  payload_not_available?: string;
 }